Moved tta types into its own file to avoid circular includes
