+++ src/runtime/TTASymbol.h
@@ -16,29 +16,26 @@
     You should have received a copy of the GNU General Public License
     along with aaltitoad.  If not, see <https://www.gnu.org/licenses/>.
  */
-#ifndef AALTITOAD_VARIABLEPREDICATE_H
-#define AALTITOAD_VARIABLEPREDICATE_H
-#include "TTASymbol.h"
-#include <extensions/overload>
+
+
+#ifndef AALTITOAD_TTASYMBOL_H
+#define AALTITOAD_TTASYMBOL_H
 #include <aaltitoadpch.h>
 
-/**
- * Contains predicates such as
- * my_external_variable > 8
- * This is used for symbolic model checking of external input variables
- * */
-struct VariablePredicate {
-    enum class PredicateComparator {
-        LESS, LESSEQ,
-        NEQ, EQ,
-        GREATER, GREATEREQ
-    };
-    std::string variable;
-    PredicateComparator comparator;
-    TTASymbol_t value;
-    static std::string ConvertToString(const PredicateComparator& cc);
-    std::string ToGuardString();
-    static VariablePredicate Merge(const VariablePredicate& a, const VariablePredicate& b);
+struct TTATimerSymbol {
+    float current_value;
 };
 
-#endif //AALTITOAD_VARIABLEPREDICATE_H
+using TTASymbol_t = std::variant<
+        int,
+        float,
+        bool,
+        TTATimerSymbol,
+        std::string
+>;
+
+TTASymbol_t TTASymbolValueFromTypeAndValueStrings(const std::string& typestr, const std::string& valuestr);
+TTASymbol_t TTASymbolTypeFromString(const std::string& typestr);
+TTASymbol_t PopulateValueFromString(const TTASymbol_t& type, const std::string& valuestr);
+
+#endif //AALTITOAD_TTASYMBOL_H
